feat(sp): adding audio and shot instancing
