--- pype/plugins/standalonepublisher/publish/extract_shot.py
@@ -1,148 +1,0 @@
-import os
-import clique
-import pype.api
-import pype.lib as plib
-
-from pprint import pformat
-
-
-class ExtractShot(pype.api.Extractor):
-    """Extract shot "mov" and "wav" files."""
-
-    label = "Extract Shot"
-    hosts = ["standalonepublisher"]
-    families = ["clip"]
-
-    # presets
-    add_representation = None  # ".jpeg"
-    add_audio = True
-
-    def process(self, instance):
-        # get ffmpet path
-        ffmpeg_path = pype.lib.get_ffmpeg_tool_path("ffmpeg")
-
-        # get staging dir
-        staging_dir = self.staging_dir(instance)
-        self.log.info("Staging dir set to: `{}`".format(staging_dir))
-
-        # Generate mov file.
-        fps = instance.data["fps"]
-        video_file_path = instance.data["editorialVideoPath"]
-        ext = os.path.splitext(os.path.basename(video_file_path))[-1]
-
-        clip_trimed_path = os.path.join(
-            staging_dir, instance.data["name"] + ext)
-
-        # check video file metadata
-        input_data = plib.ffprobe_streams(video_file_path)[0]
-        self.log.debug(f"__ input_data: `{input_data}`")
-
-        args = [
-            ffmpeg_path,
-            "-ss", str(instance.data["clipIn"] / fps),
-            "-i", video_file_path,
-            "-t", str(
-                (instance.data["clipOut"] - instance.data["clipIn"] + 1) /
-                fps
-            ),
-            "-crf", "18",
-            "-pix_fmt", "yuv420p",
-            clip_trimed_path
-        ]
-        self.log.info(f"Processing: {args}")
-        ffmpeg_args = " ".join(args)
-        output = pype.api.subprocess(ffmpeg_args)
-        self.log.info(output)
-
-        instance.data["families"].remove("review")
-        instance.data["families"].append("clip")
-        instance.data["family"] = "review"
-
-        # frame ranges
-        frame_start = int(instance.data["frameStart"])
-        frame_end = int(instance.data["frameEnd"])
-        handle_start = int(instance.data["handleStart"])
-        handle_end = int(instance.data["handleEnd"])
-
-        instance.data["representations"].append({
-            "name": ext[1:],
-            "ext": ext[1:],
-            "files": os.path.basename(clip_trimed_path),
-            "stagingDir": staging_dir,
-            "frameStart": frame_start,
-            "frameEnd": frame_end,
-            "frameStartFtrack": frame_start - handle_start,
-            "frameEndFtrack": frame_end - handle_end,
-            "fps": fps,
-            "thumbnail": True,
-            "tags": ["review", "ftrackreview", "delete"]
-        })
-
-        self.log.debug(f"Instance data: {pformat(instance.data)}")
-
-        if self.add_representation:
-            # Generate jpegs.
-            clip_img_sequence = os.path.join(
-                staging_dir, instance.data["name"] + ".%04d.jpeg"
-            )
-            args = [ffmpeg_path, "-i", clip_trimed_path, clip_img_sequence]
-            self.log.info(f"Processing: {args}")
-            output = pype.lib._subprocess(args)
-            self.log.info(output)
-
-            # collect jpeg sequence if editorial data for publish
-            # are image sequence
-            collection = clique.Collection(
-                head=instance.data["name"] + ".", tail='.jpeg', padding=4
-            )
-            for f in os.listdir(staging_dir):
-                if collection.match(f):
-                    collection.add(f)
-
-            instance.data["representations"].append({
-                "name": "jpeg",
-                "ext": "jpeg",
-                "files": list(collection),
-                "stagingDir": staging_dir
-            })
-
-        if self.add_audio:
-            audio_ext = ".wav"
-            # Generate wav file.
-            shot_wav = os.path.join(
-                staging_dir, instance.data["name"] + audio_ext)
-            # Collect data.
-            data = {}
-            for key, value in instance.data.items():
-                data[key] = value
-
-            data["family"] = "audio"
-            data["families"] = ["ftrack"]
-
-            data["subset"] = "audioMain"
-
-            data["source"] = shot_wav
-
-            data["name"] = data["subset"] + "_" + data["asset"]
-
-            data["label"] = "{} - {} - ({})".format(
-                data['asset'],
-                data["subset"],
-                audio_ext
-            )
-
-            # Create instance.
-            self.log.debug("Creating instance with: {}".format(data["name"]))
-            instance = instance.context.create_instance(**data)
-
-            args = [ffmpeg_path, "-i", clip_trimed_path, shot_wav]
-            self.log.info(f"Processing: {args}")
-            output = pype.lib._subprocess(args)
-            self.log.info(output)
-
-            instance.data["representations"] = [{
-                "name": "wav",
-                "ext": "wav",
-                "files": os.path.basename(shot_wav),
-                "stagingDir": staging_dir
-            }]
